gpu: Rename to gpu_shader_instrumentor
It is shared between GPU AV and debug printf, name should
reflect that
