+++ BUILD.gn
@@ -184,8 +184,8 @@ vvl_sources = [
   "layers/gpu/error_message/gpuav_error_message.h",
   "layers/gpu/error_message/gpuav_vuids.cpp",
   "layers/gpu/error_message/gpuav_vuids.h",
-  "layers/gpu/instrumentation/gpuav_shader_instrumentor.cpp",
-  "layers/gpu/instrumentation/gpuav_shader_instrumentor.h",
+  "layers/gpu/instrumentation/gpu_shader_instrumentor.cpp",
+  "layers/gpu/instrumentation/gpu_shader_instrumentor.h",
   "layers/gpu/resources/gpuav_descriptor_set.cpp",
   "layers/gpu/resources/gpuav_descriptor_set.h",
   "layers/gpu/resources/gpuav_subclasses.cpp",
