+++ layers/gpu/core/gpuav.h
@@ -20,7 +20,7 @@
 #include "gpu/error_message/gpuav_error_message.h"
 #include "gpu/resources/gpuav_descriptor_set.h"
 #include "gpu/resources/gpu_resources.h"
-#include "gpu/instrumentation/gpuav_shader_instrumentor.h"
+#include "gpu/instrumentation/gpu_shader_instrumentor.h"
 
 #include <unordered_map>
 #include <memory>
