+++ layers/gpu/core/gpuav_state_tracker.cpp
@@ -16,7 +16,7 @@
  */
 
 #include "gpu/core/gpuav_state_tracker.h"
-#include "gpu/instrumentation/gpuav_shader_instrumentor.h"
+#include "gpu/instrumentation/gpu_shader_instrumentor.h"
 
 namespace gpu_tracker {
