+++ layers/gpu/debug_printf/debug_printf.h
@@ -19,7 +19,7 @@
 
 #include "gpu/core/gpuav_state_tracker.h"
 #include "gpu/error_message/gpuav_error_message.h"
-#include "gpu/instrumentation/gpuav_shader_instrumentor.h"
+#include "gpu/instrumentation/gpu_shader_instrumentor.h"
 
 namespace debug_printf {
