+++ layers/gpu/instrumentation/gpu_shader_instrumentor.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "gpu/instrumentation/gpuav_shader_instrumentor.h"
+#include "gpu/instrumentation/gpu_shader_instrumentor.h"
 
 #include "gpu/core/gpuav_state_tracker.h"
 #include "chassis/chassis_modification_state.h"
